+++ ui/src/components/Editor/Scratch.tsx
@@ -13,7 +13,6 @@ import { FlexColumnSpaced } from '../View/FlexColumnSpaced'
 import {
   getTypedHolesFromEditor,
   getSourceItemsFromEditor,
-  getErrorLocations,
   getErrorLocationsFromEditor,
 } from '../../reducer/editor/selector'
